requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.1.2
pymongo==4.5.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
pytz==2023.3
